ATS/requirements.txt
streamlit
PyPDF2
google-generativeai
pdfplumber
openai
time